logf -> log

+++ src/kernels/cl/mandelbrot.cl
@@ -41,7 +41,7 @@ __kernel void mandelbrot(__global float* results,
 
     float result = iter;
     if (isSmoothing && iter != iters) {
-        result = result - logf(logf(sqrtf(x * x + y * y)) / logf(threshold)) / logf(2.0f);
+        result = result - log(log(sqrt(x * x + y * y)) / log(threshold)) / log(2.0f);
     }
 
     result = 1.0f * result / iters;
